created exception DTO

+++ src/main/java/com/example/FakeStore/dtos/fakeStoreDTO.java
@@ -6,15 +6,15 @@ import lombok.Setter;
 @Setter
 public class fakeStoreDTO{
 
-    public Long id;
+    public int id;
     public String title;
     public double price;
     public String category;
     public String description;
     public String image;
 
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; }
 
     public String getTitle() { return title; }
     public void setTitle(String title) { this.title = title; }
